+++ lisp/cov2.lisp
@@ -1,3 +1,5 @@
+;;;; Sample coverage file
+
 (("/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test.lisp" "00c69412bf934617158b59331dbca964db71d7aec8e2f28bb553e48d21239ec6"
   "/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test2.lisp" "55458ffb66e84d015f9e034214cbdada1a8f4d6024a1c8904d1306100f92db69")
  ("/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test2.lisp" 1 1 1 14 1)