+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added then slipr application. slipr is used with shebang (#!) to run
   lisp code by adding `#!/usr/bin/env slipr` as the first line in a
   .lisp file.
+- Added cover.lisp that displays coverage data from a save coverage file in the lisp directory.
 
 ### Fixed
 - The CompileList function now correctly handles package prefixes.